*** Add support for returning the end-of-parsing position in toInt & family
If endpos is nullptr, we behave like we did before: the entire string
must be parsed, or we'll fail. If it isn't nullptr, we stop at the first
character we can't parse and store its position there.

This is implemented for QString and QByteArray only. We can't implement
it for QLocale because of the parsing of group (thousand) separators:
since we strip characters from the middle of the digit sequence, we
can't get the endptr from qstrto{d,ll,ull}. We'd have to either keep an
expensive index mapping or rewrite the qstrto{d,ll,ull} code to skip the
thousand separators too.

Task-number: QTBUG-66115
Change-Id: Ic50b26360e404f31271cecb90fcb160c0386eaa2