*** Support reporting non-numeric suffixes when parsing numbers
If convertedChars is nullptr, we behave like we did before: the entire
string must be parsed, or we'll fail. If it isn't nullptr, we stop at
the first character we can't parse and store its index there.

This is implemented for QString and QByteArray only. We can't implement
it for QLocale because of the parsing of group (thousand) separators:
since we strip characters from the middle of the digit sequence, we
can't get the end postion from qstrto{d,ll,ull}. We'd have to either
keep an expensive index mapping or rewrite the qstrto{d,ll,ull} code
to skip the thousand separators too.

[ChangeLog][QtCore][QString][QByteArray] Added overloads to get the
number of converted characters from toInt(), toDouble() and related
functions.

Done-with: Andre Hartmann <aha_1980@gmx.de>
Task-number: QTBUG-66115
Change-Id: Ic50b26360e404f31271cecb90fcb160c0386eaa2